Added Github Source and Source Interface
A source is where the updates come from. Since there can be many types
of sources there’s an interface that can be extended to create new ones.

The first source is Github, and uses the Releases and Assets API to
identify each version that can be used to update from.